# Add Ascend NPU accelerator support