--- a/src/accelerate/utils/imports.py
+++ b/src/accelerate/utils/imports.py
@@ -103,6 +103,8 @@ def is_bf16_available(ignore_tpu=False):
         return not ignore_tpu
     if torch.cuda.is_available():
         return torch.cuda.is_bf16_supported()
+    if is_npu_available():
+        return False
     return True