--- a/src/accelerate/utils/operations.py
+++ b/src/accelerate/utils/operations.py
@@ -287,6 +287,8 @@ def gather(tensor):
         return _tpu_gather(tensor)
     elif PartialState().distributed_type in CUDA_DISTRIBUTED_TYPES:
         return _gpu_gather(tensor)
+    elif PartialState().distributed_type in DistributedType.MULTI_NPU:
+        return _gpu_gather(tensor)
     elif PartialState().distributed_type in DistributedType.MULTI_XPU:
         return _gpu_gather(tensor)
     elif PartialState().distributed_type == DistributedType.MULTI_CPU:
@@ -319,6 +321,8 @@ def gather_object(object: Any):
     if PartialState().distributed_type == DistributedType.TPU:
         raise NotImplementedError("gather objects in TPU is not supported")
     elif PartialState().distributed_type in CUDA_DISTRIBUTED_TYPES:
+        return _gpu_gather_object(object)
+    elif PartialState().distributed_type in DistributedType.MULTI_NPU:
         return _gpu_gather_object(object)
     elif PartialState().distributed_type in DistributedType.MULTI_XPU:
         return _gpu_gather_object(object)
@@ -361,6 +365,8 @@ def broadcast(tensor, from_process: int = 0):
         return _tpu_broadcast(tensor, src=from_process, name="accelerate.utils.broadcast")
     elif PartialState().distributed_type in CUDA_DISTRIBUTED_TYPES:
         return _gpu_broadcast(tensor, src=from_process)
+    elif PartialState().distributed_type in DistributedType.MULTI_NPU:
+        return _gpu_gather_object(object)
     elif PartialState().distributed_type in DistributedType.MULTI_XPU:
         return _gpu_broadcast(tensor, src=from_process)
     elif PartialState().distributed_type == DistributedType.MULTI_CPU:
@@ -386,6 +392,8 @@ def broadcast_object_list(object_list, from_process: int = 0):
         for i, obj in enumerate(object_list):
             object_list[i] = xm.mesh_reduce("accelerate.utils.broadcast_object_list", obj, lambda x: x[from_process])
     elif PartialState().distributed_type in CUDA_DISTRIBUTED_TYPES:
+        torch.distributed.broadcast_object_list(object_list, src=from_process)
+    elif PartialState().distributed_type in DistributedType.MULTI_NPU:
         torch.distributed.broadcast_object_list(object_list, src=from_process)
     elif PartialState().distributed_type in DistributedType.MULTI_XPU:
         torch.distributed.broadcast_object_list(object_list, src=from_process)
@@ -505,6 +513,8 @@ def reduce(tensor, reduction="mean"):
         if state.distributed_type == DistributedType.TPU:
             xm.all_reduce("sum", cloned_tensor)
         elif state.distributed_type.value in CUDA_DISTRIBUTED_TYPES:
+            torch.distributed.all_reduce(cloned_tensor, ReduceOp.SUM)
+        elif state.distributed_type.value in DistributedType.MULTI_NPU:
             torch.distributed.all_reduce(cloned_tensor, ReduceOp.SUM)
         elif state.distributed_type.value in DistributedType.MULTI_XPU:
             torch.distributed.all_reduce(cloned_tensor, ReduceOp.SUM)