@HuggingFaceDocBuilderDev 2023-07-04
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/accelerate/pr_1676). All of your documentation changes will be reflected on that endpoint.